split display_input, action_visuals, table part from screen.go
